Checkpoint 02-08-2023
A lot of stuff has changed since the last commit.

Most of the commands have been ported from the pre-Haven version of the game.

The scripting engine has been updated to include all sorts of things including an audio API.

The game's UI has been updated to respond to changing window dimensions and therefore looks a bit different. (This took the longest)

There is now a simple notification system in the form of the NotifyLayer in UI/NotifyLayer.cs. There are also some new sound effects in the Audio/SFX folder.

Currently working on the NETChat layer and getting the singleplayer friends and narrative chat thing working. It will all be powered by a separate script engine instance with its own API and support multiple branching paths in dialogue when it is done.